Adds QR code modal for donation details
Implements a modal to display a larger view of donation QR codes.
This allows users to easily view and scan the QR codes for donations.
The modal is triggered by clicking on the QR code image, and it displays the image along with the donation name.
Also adds a hover effect to the QR code image to indicate interactivity.
</body